Minor Adaptions
git-svn-id: https://svn.code.sf.net/p/quex/code/trunk@4605 db9deaaf-66c7-40b0-811a-ad7aae109365

+++ quex/code_base/buffer/TESTS/navigation/TEST/basic_functionality.c
@@ -240,7 +240,7 @@ find_reference(const char* file_stem)
         snprintf(&file_name[0], 255, "%s.dat", file_stem);
     }
     else {
-        snprintf(&file_name[0], 255, "%s-%i-%s.dat", file_stem, sizeof(QUEX_TYPE_LEXATOM)*8, 
+        snprintf(&file_name[0], 255, "%s-%i-%s.dat", file_stem, (int)sizeof(QUEX_TYPE_LEXATOM)*8, 
                  QUEXED(system_is_little_endian)() ? "le" : "be");
     }
     return &file_name[0];
