+++ quex/code_base/buffer/TESTS/navigation/TEST/test-Converter.c
@@ -54,7 +54,7 @@ main(int argc, char** argv)
 
     if( argc > 1 && strcmp(argv[1], "--hwut-info") == 0 ) {
         printf("Buffer Tell&Seek: LexatomLoader_Converter_IConv (BPC=%i, FALLBACK=%i);\n", 
-               BPC, QUEX_SETTING_BUFFER_MIN_FALLBACK_N);
+               (int)BPC, (int)QUEX_SETTING_BUFFER_MIN_FALLBACK_N);
         printf("CHOICES: ICU-linear, ICU-stepping,\n"
                "         IConv-linear, IConv-stepping,\n"
                "         IConv-stepping-cls, ICU-stepping-cls;\n"
